VL6/requirements.txt
ipykernel
pillow
datasets
torch
transformers
accelerate
pydantic
PyYAML
jinja2==3.1.0